reworked solution structure to acommodate additional projects
