+++ IpfsApi.sln
@@ -13,9 +13,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Documentation", "doc\Documentation.csproj", "{F3A32EA9-0B2F-46A3-B47A-33B4C04BD423}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IpfsApi", "src\IpfsApi.csproj", "{F3C81C57-C283-4E07-B765-DEABCFB22136}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IpfsApiTests", "test\IpfsApiTests.csproj", "{B459FBC7-4A28-4170-AD83-7348A403407F}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IpfsApi", "src\IpfsApi\IpfsApi.csproj", "{45CF530D-685E-4D27-9570-9B729ABA7FD1}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -25,14 +25,14 @@ Global
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{F3A32EA9-0B2F-46A3-B47A-33B4C04BD423}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{F3A32EA9-0B2F-46A3-B47A-33B4C04BD423}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F3C81C57-C283-4E07-B765-DEABCFB22136}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F3C81C57-C283-4E07-B765-DEABCFB22136}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F3C81C57-C283-4E07-B765-DEABCFB22136}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F3C81C57-C283-4E07-B765-DEABCFB22136}.Release|Any CPU.Build.0 = Release|Any CPU
 		{B459FBC7-4A28-4170-AD83-7348A403407F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{B459FBC7-4A28-4170-AD83-7348A403407F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B459FBC7-4A28-4170-AD83-7348A403407F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B459FBC7-4A28-4170-AD83-7348A403407F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{45CF530D-685E-4D27-9570-9B729ABA7FD1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{45CF530D-685E-4D27-9570-9B729ABA7FD1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{45CF530D-685E-4D27-9570-9B729ABA7FD1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{45CF530D-685E-4D27-9570-9B729ABA7FD1}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
