+++ src/IpfsApi/CoreApi/DnsApi.cs
@@ -1,0 +1,34 @@
+﻿using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Net.Http;
+using System.Text;
+using System.Threading;
+using System.Threading.Tasks;
+using Ipfs.CoreApi;
+using System.IO;
+
+namespace Ipfs.Api
+{
+
+    class DnsApi : IDnsApi
+    {
+        IpfsClient ipfs;
+
+        internal DnsApi(IpfsClient ipfs)
+        {
+            this.ipfs = ipfs;
+        }
+
+        public async Task<string> ResolveAsync(string name, bool recursive = false, CancellationToken cancel = default(CancellationToken))
+        {
+            var json = await ipfs.DoCommandAsync("dns", cancel,
+                name,
+                $"recursive={recursive.ToString().ToLowerInvariant()}");
+            var path = (string)(JObject.Parse(json)["Path"]);
+            return path;
+        }
+    }
+}
