+++ src/IpfsApi/CoreApi/GenericApi.cs
