+++ test/IpfsApiTests.csproj
@@ -15,7 +15,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\src\IpfsApi.csproj" />
+    <ProjectReference Include="..\src\IpfsApi\IpfsApi.csproj" />
   </ItemGroup>
