+++ test/TestFixture.cs
@@ -15,6 +15,6 @@ namespace Ipfs.Api
         /// </summary>
         //IpfsClient.DefaultApiUri = new Uri("http://127.0.0.1.:5001");
 
-        public static IpfsClient Ipfs = new IpfsClient();
+        public static IpfsClient Ipfs => new IpfsClient();
     }
 }
